Redesign of the `allow::Ref`-based API.
This is a rethinking of the design. With this design, a console write command would have the protoype:

```
fn console_write(to_write: AllowRef<&[u8]>) -> Result<usize, ErrorCode>;
```

An `AllowRef` can only exist if the buffer if something guarantees that the buffer it points to will be unshared before being dropped. There are multiple ways to construct an `AllowRef`:

1. From a `'static` reference, where the buffer will never be dropped.
2. Form a `Pin<&mut Buffer>`, where `Buffer::drop` unshares the buffer.

This new design is currently incomplete, and I'm not sure what the size impact will be.